no more blockmap limits? yay

+++ latest/P_LOCAL.H
@@ -311,14 +311,14 @@ P_RadiusAttack
 // P_SETUP
 //
 extern byte*            rejectmatrix;   // for fast sight rejection
-extern short*           blockmaplump;   // offsets in blockmap are from here
-extern short*           blockmap;
-extern int              bmapwidth;
-extern int              bmapheight;     // in mapblocks
+extern int*           blockmaplump;   // offsets in blockmap are from here
+extern int*           blockmap;
+extern int            bmapwidth;
+extern int            bmapheight;     // in mapblocks
 extern fixed_t          bmaporgx;
 extern fixed_t          bmaporgy;       // origin of block map
 extern mobj_t**         blocklinks;     // for thing chains
-extern boolean			level_has_bosses;
+extern boolean			level_has_bosses; // dumb nozomi analog check
 
 
 //
